move around TalkBack and try to fix ask_address_question_ bug
